+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+Application.ensure_all_started(:credo)
 ExUnit.start()
+Code.put_compiler_option(:warnings_as_errors, true)