+++ CHANGELOG.md
@@ -2,5 +2,18 @@
 
 ## Unreleased
 
+- Added checks
+  - BlitzCredoChecks.DocsBeforeSpecs
+  - BlitzCredoChecks.DoctestIndent
+  - BlitzCredoChecks.LowercaseTestNames
+  - BlitzCredoChecks.NoAsyncFalse
+  - BlitzCredoChecks.NoDSLParentheses
+  - BlitzCredoChecks.NoIsBitstring
+  - BlitzCredoChecks.SetWarningsAsErrorsInTest
+  - BlitzCredoChecks.StrictComparison
+  - BlitzCredoChecks.UseStream
+- Document use of library
+
 ## Created Repo (2022-04-07)
+
 - Created the project!